Rwjs/remove delete filelisthandler (#125)
* Removed the deletion of the filelisthandler from OnDisconnect. The logic for doing this
did not fit in with the reference counting introduced in the last merge, unless code
was changed elsewhere. That could be done but since there is only ever one filelisthandler
there does not seem to be any point in deleting it.

Using the same logic, moved the constructor for the filelisthandler to the main function
since if there is only ever one, why not just make it at startup.

* Moved the FileListHandler constructer to after the permissions are read. This makes no
difference to the execution of the program, but is more logical.